Multicast related fixes.
Added joining a membership for servers, which causes the igmp
query to be sent on the given interface. Fixed the TTL setting
for multicast clients to IP_MULTICAST_TTL.

Added joining source specific groups with the -M <IP> switch.

Changed the start and end IP calculation to use direct IPs for
IPv4. This allows for greater range. This prompted the adding of
-U for increasing the number of open files to the OS max.